requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
aiofiles==23.2.1
databricks-sdk>=0.40.0
sqlglot>=26.0.0
